requirements.txt
recbole==1.0.0
torch>=1.7.0
numpy==1.20.3
scipy==1.6.0
hyperopt>=0.2.4
tqdm>=4.48.2
scikit_learn>=0.23.2
pyyaml>=5.1.0
colorlog==4.7.2
colorama==0.4.4
tensorboard>=2.5.0
pandas
xgboost
torch_geometric
